Reorganization and changes to make setup smoother
* Removing items from base __init__ file so they are not imported when
trying to get to version.
* Better setup.
