+++ .travis.yml
@@ -18,5 +18,6 @@ deploy:
   username: panoptes
   password:
     secure: x4qLLV0mdoXfp12CDjFLaK0AWBxp7K6UfD+3TGgaVpYJbu2hgRDDi9C4iSbS2pHb6NDumiw5viOxpcj9oZnPlIczvC83YoVOqgzSS3gmzpyh0hSsHkZvWOPrTajUV/VqFrGdS3anN1hUEvLrnfQXeZ6+u2jomlbLRueCAAjSkn0dT1XvvoL5ObJUjR/VPZKgVoJmYuWhUYCYBxYxlGEPeRq6jzwlECQduvyk+ZltdLnnlXQT9yQDDrXwN19GweYfnpUwEOlBpVzX9rtiOG+SyZGVRhDzrWaGWWVOGwdCMhiEv+UR69v8cIg4BGTTDEt1tLa050vj8SCTh8c7wY5hC/n9VzDEOROq/pwVrZE0MPEt646ZB0j6+oZ4j6DsHkyMx2V4G/YoqTwnV1zyG3eQL0YU27vtg3TVJYu3Y07zbjaCCJnlzRuRFknXuwqwWkpYl54HlNHL31UNXhN9p4L6WFEWoi6U5MmAwzHK4tWAGAoCgBIGy6MSkX7iLxkLV6IRgvLQntxjppqd3otriuAazntquZwYNgNvU5L0GPAh1icR9Wt8B8lrQ+2V+HuqEEUbOrrgdoa9b06+M1PYg+PvB93s9SSm3Nb/3dShWehWCMbIh3UfwHyJbKPIRcNPZ1Hgx2gx9LUhs0KDws6NswDBogAImafjL5u1zhT4NnNQMuE=
+  distributions: sdist bdist_wheel    
   on:
-    branch: master
+    branch: develop
