+++ panoptes_utils/version.py
@@ -2,4 +2,4 @@ major = 0
 minor = 0
 patch = 6
 
-__version__ = '{}.{}.{}'.format(major, minor, patch)
+__version__ = '{}.{}.{}-dev'.format(major, minor, patch)
